# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/legendaryPrompt": {
            "post": {
                "description": "Post prompt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prompts"
                ],
                "summary": "PingPong",
                "parameters": [
                    {
                        "description": "prompt",
                        "name": "prompt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Prompt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/prompt": {
            "post": {
                "description": "Post prompt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prompts"
                ],
                "summary": "PostPrompt",
                "parameters": [
                    {
                        "description": "prompt",
                        "name": "prompt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Prompt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "is_ai": {
                    "type": "boolean"
                }
            }
        },
        "types.Prompt": {
            "type": "object",
            "properties": {
                "act_id": {
                    "type": "integer"
                },
                "msg": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Message"
                    }
                }
            }
        }
    }
}